esp/lib/django-fab-deploy-ubuntu12-pgmods/requirements.txt
jinja2
fabric >= 1.4.0
fabric-taskset